+++ src/screens/Repository/controllers/animationsController.js
