+++ test/screens/Repository/controllers/messagesController.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 const {
   createMessage,
-} = require("../src/screens/Repository/messagesController.js");
+} = require("../../../../src/screens/Repository/controllers/messagesController.js");
 
 describe("createMessage", () => {
   it("should create a message element with the given message and author", () => {
